// packages/src/animation/types.ts
//

export type QTAnimatorId = string


export interface QTAnimationInterpolator {
  type: QTAnimationInterpolatorType
}


export enum QTAnimationInterpolatorType {
  QT_ACCELERATE_DECELERATE_INTERPOLATOR = 1,
  QT_ACCELERATE_INTERPOLATOR = 2,
  QT_ANTICIPATE_INTERPOLATOR = 3,
  QT_ANTICIPATE_OVERSHOOT_INTERPOLATOR = 4,
  QT_BOUNCE_INTERPOLATOR = 5,
  QT_CYCLE_INTERPOLATOR = 6,
  QT_DECELERATE_INTERPOLATOR = 7,
  QT_LINEAR_INTERPOLATOR = 8,
  QT_OVERSHOOT_INTERPOLATOR = 9,
  QT_FAST_OUT_LINEAR_IN_INTERPOLATOR = 10,
  QT_FAST_OUT_SLOW_IN_INTERPOLATOR = 11,
  QT_PATH_INTERPOLATOR = 12,
}

export enum QTAnimationPropertyName {
  QT_ANIMATION_PROPERTY_NAME_ALPHA = 'alpha',

  QT_ANIMATION_PROPERTY_NAME_ROTATION = 'rotation',
  QT_ANIMATION_PROPERTY_NAME_ROTATION_X = 'rotationX',
  QT_ANIMATION_PROPERTY_NAME_ROTATION_Y = 'rotationY',
  QT_ANIMATION_PROPERTY_NAME_ROTATION_Z = 'rotationZ',

  QT_ANIMATION_PROPERTY_NAME_SCALE_X = 'scaleX',
  QT_ANIMATION_PROPERTY_NAME_SCALE_Y = 'scaleY',
  QT_ANIMATION_PROPERTY_NAME_SCALE_Z = 'scaleZ',

  QT_ANIMATION_PROPERTY_NAME_TRANSLATION_X = 'translationX',
  QT_ANIMATION_PROPERTY_NAME_TRANSLATION_Y = 'translationY',
  QT_ANIMATION_PROPERTY_NAME_TRANSLATION_Z = 'translationZ',
}


export enum QTAnimationRepeatMode {
  QT_ANIMATION_REPEAT_MODE_RESTART = 1,
  QT_ANIMATION_REPEAT_MODE_REVERSE = 2,
  QT_ANIMATION_REPEAT_MODE_INFINITE = -1,
}

export enum QTAnimationValueType {
  QT_ANIMATION_VALUE_TYPE_INT = 'ofInt',
  QT_ANIMATION_VALUE_TYPE_FLOAT = 'ofFloat'
}